% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{clear.ontology.mappings}
\alias{clear.ontology.mappings}
\title{Clear All Registered Ontology Mappings}
\usage{
clear.ontology.mappings()
}
\description{
Clear All Registered Ontology Mappings
}
\details{
Clears all registered ontology mappings, such that new mappings can be added
}
