% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{find.ontology.mapping}
\alias{find.ontology.mapping}
\title{Find Ontology Mapping}
\usage{
find.ontology.mapping(
  from.ontology,
  to.ontology,
  required.dimensions,
  required.dim.names,
  get.two.way.alignment,
  allow.non.overlapping.incomplete.dimensions,
  from.ontology.name = ifelse(get.two.way.alignment, "ontology.1", "the from.ontology"),
  to.ontology.name = ifelse(get.two.way.alignment, "ontology.2", "the to.ontology"),
  get.details = F
)
}
\arguments{
\item{from.ontology, to.ontology}{The ontologies of the from and to arrays}

\item{required.dimensions}{The dimensions that should be present in the aligned product}

\item{required.dim.names}{Any specific dimension values that should be present in the aligned product}

\item{get.two.way.alignment}{Logical indicating whether we need one mapping, that maps from.dim.names such that}

\item{allow.non.overlapping.incomplete.dimensions}{Logical indicating whether, when we map to an incomplete dimension, there can be no overlap in the from and to values for that dimension}
}
\value{
If get.two.way.alignment is false, returns a list of one or more ontology.mapping objects,
         such that applying these mappings to an array with from.dim.names ('from.arr')
         would yield an array with dimnames ('to.dim.names')
         
       If get.two.way.alignment is true, returns a list with two elements,
         each of which is itself a list of one or more ontology.mapping objects, such that
         applying the first set of mappings to an array with from.dimnames
         and applying the second set of mappings to an array with to.dimnames
         would both yield arrays with dimensions 'align.on.dimensions', both with
         the same dimnames
}
\description{
Find Ontology Mapping
}
