% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_array_helpers.R
\name{get.collapse.array.indices}
\alias{get.collapse.array.indices}
\title{Get a Set of Indices to Collapse One Array into Another}
\usage{
get.collapse.array.indices(large.arr.dim.names, small.arr.dim.names)
}
\value{
A list with four elements: $large.indices, $small.indices, $large.n, $small.n
 large.indices and small.indices are integer vectors of equal length, such that the value at large.indices[i] goes into the value at small.indices[i] if collapsing
}
\description{
Get a Set of Indices to Collapse One Array into Another
}
