% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{get.ontology.mapping.matrix}
\alias{get.ontology.mapping.matrix}
\title{Get a Matrix that Applies an Ontology Mapping}
\usage{
get.ontology.mapping.matrix(
  ontology.mapping,
  from.dim.names,
  to.dim.names = NULL,
  error.prefix = ""
)
}
\arguments{
\item{ontology.mapping}{An object of class 'ontology.mapping' - obtained either through \link{register.ontology.mapping}, \link{get.ontology.mapping}, \link{get.mappings.to.align.ontologies}}

\item{to.dim.names}{The dimnames that the resulting transformed data should have (can represent a subset of the dimnames that the ontology mapping knows how to transform). If NULL, the ontology mapping will keep the dimnames in from.arr except for dimensions that are transformed}

\item{error.prefix}{A character value to be prepended to any error messages thrown by the function}
}
\description{
Get a Matrix that Applies an Ontology Mapping
}
