% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{get.ontology.mapping}
\alias{get.ontology.mapping}
\title{Get an Ontology Mapping to Transform Data}
\usage{
get.ontology.mapping(
  from.ontology,
  to.ontology,
  allow.non.overlapping.incomplete.dimensions = F
)
}
\arguments{
\item{from.ontology}{The ontology to which the data to be transformed conform. Must be an 'ontology' object as created by \code{\link{ontology}}}

\item{to.ontology}{The ontology to which the transformed data should conform. Must be an 'ontology' object as created by \code{\link{ontology}}}

\item{allow.non.overlapping.incomplete.dimensions}{Logical indicating whether, when we map to an incomplete dimension, there can be no overlap in the from and to values for that dimension}
}
\value{
Either an object of class 'ontology.mapping' or NULL if no mapping that would bridge the differences is found
}
\description{
Get an Ontology Mapping to Transform Data
}
