% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{register.calibrated.parameters.for.version}
\alias{register.calibrated.parameters.for.version}
\title{Register Parameters to Be Used in Calibrating a JHEEM Version}
\usage{
register.calibrated.parameters.for.version(
  version,
  distribution,
  sampling.blocks,
  apply.function,
  join.with.previous.version,
  calibrate.to.year
)
}
\arguments{
\item{version}{The name of an EHE version for which a specification has been registered with \code{\link{register.model.specification}}}

\item{distribution}{An object of class 'distribution' that specifies the prior distribution for the parameters}

\item{sampling.blocks}{A list (which may optionally be named) of character vectors. Each element (a character vector) denotes a subset of the parameters to be sampled as a single block during the calibration process. These elements must contain ONLY values present in distribution@var.names, and every value in distribution@var.names must be present in at least one of the blocks}

\item{apply.function}{A function that knows how to apply parameter values to a jheem.model.settings object. The function should take two arguments: (1) 'model.settings', an object of class jheem.model.settings, and (2) 'parameters', a named numeric vector}

\item{join.with.previous.version}{Whether the distribution, apply.function, and sampling blocks should be merged with those of the previous version}

\item{calibrate.to.year}{The year to which the calibration should be run before sampling parameters}
}
\description{
Register Parameters to Be Used in Calibrating a JHEEM Version
}
\details{
*Calibrated* parameters have their values fitted though a Bayesian calibration process
}
