% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_age_year_helpers.R
\name{restratify.age.counts}
\alias{restratify.age.counts}
\title{Re-distribute counts according to one set of age brackets into a different set of age brackets}
\usage{
restratify.age.counts(
  counts,
  desired.age.brackets,
  given.age.brackets = NULL,
  smooth.infinite.age.to = Inf,
  allow.extrapolation = F,
  na.rm = F,
  method = c("pclm", "monoH.FC", "hyman")[2],
  error.prefix = ""
)
}
\arguments{
\item{counts}{Either (a) a dimensionless numeric vector, or (b) a numeric array, with  one dimension named "age". The "given" age brackets according to which these counts are distributed must be continuous - with no gaps between them}

\item{desired.age.brackets, given.age.brackets}{The age brackets according to which the ages in counts are given or desired to be restratified to. Either (1) a character vector giving names of brackets, with one name for each age bracket value or (b) a numeric vector of age endpoints, with one more element than the number of age bracket values. given.age.brackets should be present only if names(counts) is absent (if counts is a vector) or if dimnames(counts)$age is null (if counts is an array)}

\item{smooth.infinite.age.to}{In building a spline over counts, what value should an infinite bound in an age range be replaced with}

\item{allow.extrapolation}{A logical value indicating whether extrapolation should be allowed in restratifying. Extrapolating is required when one of the desired age brackets includes ages not present in any of the given age brackets}

\item{na.rm}{Whether NAs should be ignored in counts}

\item{method}{The method to be passed to the smoother. Options are 'pclm' (default, uses Penalized Composite Link Model), 'monoH.FC' (Fritsch-Carlson monotonic splines), or 'hyman' (Hyman's monotonic splines)}

\item{error.prefix}{A character value to be prepended to any error messages}
}
\description{
Re-distribute counts according to one set of age brackets into a different set of age brackets
}
