% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{version.has.been.registered}
\alias{version.has.been.registered}
\title{Check Whether a JHEEM Model Specification has been Registered for a Version}
\usage{
version.has.been.registered(version)
}
\arguments{
\item{version}{A character vector with one or more version names}
}
\value{
A logical vector of the same length as 'version'
}
\description{
Check Whether a JHEEM Model Specification has been Registered for a Version
}
\seealso{
Other JHEEM Version Management Functions: 
\code{\link{get.specification.for.version}()},
\code{\link{register.model.specification}()}
}
\concept{JHEEM Version Management Functions}
